requirements.txt
scipy
scikit-learn
torchmetrics